Clean up TODO and FIXME tasks [ch10755]
The main change was reworking the update remote process to never
merge the metadata produced with a previous version.

The commit message was moved to a task, to enable easy configuration.
That comes at the price of not having the two metadata files easy at
hand.

Removed some cruft used early on for testing, and improved database
acquisition.