Use local metadata
On git fetch db task, discard metadata repository information about the project, and use the project metadata instead.

This removes the need to update self (that is, committing a change to the metadata repository) before computing the
build topology, which, in turn, allows tasks that need the build topology but must not update self.