Add trickle{Outdated,Update,UpdateSession}Dependencies
trickleOutdatedDependencies will return a repository's own outdated dependencies.

trickleUpdateSessionDependencies will update the current session libraryDependencies. It's an aggregated task,
so it can be run per module. Its changes can be "session save"'d, though that is not recommended.

trickleUpdateDependencies should be overridden to do the right thing. It's not aggregated and should use the
output of trickleOutdatedDependencies to perform its task. It defaults to trickleUpdateSessionDependencies,
but since that doesn't save the result, it is a no-op with regards to file changes.