Add trickleUpdatedDependencies, for metadata information
The trickleOutdatedDependencies task reports the dependencies which
have newer version than what's currently configured in the project.

The new trickleUpdatedDependencies task reports, instead, what
dependencies have been updated since the last update of the current
project in the metadata repository.

The intuition is that, if you run trickleUpdateDependencies, then
trickleOutdated dependencies will be empty, and
trickleUpdatedDependencies will display all changes that will be sent
to the metadata repository if you run trickleUpdateSelf.